requirements.txt
requests==2.32.5
beautifulsoup4==4.14.2
lxml==6.0.2
python-telegram-bot==22.5